add product type save and update

+++ src/main/java/mil/army/moda/pao_dashboard/product_types/ProductTypeRepository.java
@@ -2,6 +2,7 @@ package mil.army.moda.pao_dashboard.product_types;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductTypeRepository extends JpaRepository<ProductType, Integer> {
+public interface ProductTypeRepository extends JpaRepository<ProductType, Long> {
+
 
 }
